454: Small refactor
